clean up actions
