+++ client/app/reader/actions.js
@@ -206,13 +206,6 @@ export const showPlaceAnnotationIcon = (pageIndex, x, y) => ({
   }
 });
 
-export const movePlacingAnnotation = (direction) => ({
-  type: Constants.MOVE_PLACE_ANNOTATION_ICON,
-  payload: {
-    direction
-  }
-});
-
 export const hidePlaceAnnotationIcon = () => ({ type: Constants.HIDE_PLACE_ANNOTATION_ICON });
 
 export const placeAnnotation = (pageNumber, coordinates, documentId) => ({
